Remove headers from LOCAL_SRC_FILES
The build system does not use headers in LOCAL_SRC_FILES. To prevent
typos, it's becoming a warning to add something to LOCAL_SRC_FILES that
is not used.

Removes LOCAL_C_INCLUDES for lz4, which now exports its include paths
automatically.

Removes LOCAL_MODULE_TAGS := optional, which is the default.

Change-Id: Ibc957290967731d7a0d099d12e69216d8595268c