cheets: fix squashfs-tools/android.c to support '-mount-point /'
When mksquashfs is called with '-android-fs-config', which is the
default Android.mk behavior, mksquashfs ignores the original file
permissions in out/target/product/<name>/ and tries to use the
permission returned from fs_config() in libcutils instead. The
libcutils function has a pattern matching table, and returns a
permission like "root/root, 0644" based on the table.

However, when '-android-fs-config' is used with '-mount-point /',
mksquashfs calls fs_config with a bad path name like '//system/foo'
which fs_config does not recognize. This CL fixes the issue so
that cheets can compress the '/' image with mksquashfs.

BUG=25115863
TEST=Ran 'unsquashfs -ll system.raw.img' and confirmed executables
 have u+x permissions. Without the fix, they don't have u+x.
Change-Id: Iabb23af788fee85218298ea1551d067e11a8567d