Fix rounding error when generating .map files.
This patch fixes the rounding of bytes to blocks in the .map file
generation process so all blocks are assigned to files while still
keeping the files not overlaping. This fixes the previous behavior
where the last 4K-block of a file wouldn't be assigned to that block,
even in the default 4k-align case leaving the last block of each file
unassigned.

To achieve this, we assign each block based on the first byte of the
block, which prevents overlapping and assigns all the blocks.

Bug: 28150981
Test: `make dist` compared system.map with the previous version. No more "gaps" observed.
Change-Id: Ib1a77036a578725a25bafa28b4f86d9dc859adb2