Fix rounding error when generating .map files.
This patch fixes the rounding of bytes to blocks in the .map file
generation process so all blocks are assigned to files while still
keeping the files not overlaping. This fixes the previous behavior
where the last 4K-block of a file wouldn't be assigned to that block,
even in the default 4k-align case leaving the last block of each file
unassigned.

To achieve this, we assign each block based on the first byte of the
block, which prevents overlapping and assigns all the blocks.

Bug: 28150981
Test: `make dist` compared system.map with the previous version. No more "gaps" observed.

(cherry picked from commit 4445cd175ee8ed20fdec0537a76a6e3cb65fdb53)

Change-Id: Iad20a7657f023ef5d2c63fad51e695d86671ff2f